Fix Windows build break
Fixing:

resource\csdk\connectivity\src\adapter_util\ca_adapter_net_ssl.c(1345):
error C2220: warning treated as error - no 'object' file generated

resource\csdk\connectivity\src\adapter_util\ca_adapter_net_ssl.c(1345):
warning C4389: '!=': signed/unsigned mismatch

Reviewed-on: https://gerrit.iotivity.org/gerrit/20055

+++ resource/csdk/connectivity/src/adapter_util/ca_adapter_net_ssl.c
@@ -1341,8 +1341,8 @@ static void SetupCipher(mbedtls_ssl_config * config, CATransportAdapter_t adapte
     {
         for (unsigned int i = 0; i < SSL_CIPHER_MAX - 1; i++)
         {
-            if (MBEDTLS_TLS_ECDHE_PSK_WITH_AES_128_CBC_SHA256 != tlsCipher[i][0] &&
-                    i != g_caSslContext->cipher)
+            if ((MBEDTLS_TLS_ECDHE_PSK_WITH_AES_128_CBC_SHA256 != tlsCipher[i][0]) &&
+                    (i != (unsigned int)g_caSslContext->cipher))
             {
                 g_cipherSuitesList[index] = tlsCipher[i][0];
                 index ++;
